These are the games sanke game packman game tictactoe game as well.
The concept of Pac-Man is quite simple. Pac-Man eats dots in a maze to score points. Avoid the ghosts unless you have just eaten a power-up, in which case ghosts are tasty. In this series we have gradually introduced new elements of Pygame Zero and also concepts around writing games. This is the first instalment in a two-part tutorial which will show you some more tricks to writing arcade games with Pygame Zero. We will also use some more advanced programming concepts to make our games even better. 

Snake Game. It’s the most famous game we used to play in our childhood before the advent of smartphones. This is a very...
Python Snake Game Project. The objective of this project is to implement the snake game using Python. It requires a...
Project Prerequisites. This python project requires a good knowledge of the pygame module, concept of functions, loops...

Step 1: Tic-tac-toe Design. We will be playing Tic-tac-toe on the command line, therefore, the first thing we have to do...
Step 2: Store information using data structures. The core of any game is the game mechanics behind it. Since this is a...
Step 3: Game Loop. Every game, has some kind of game loop, which runs until some player wins or the game ends in a draw.
Step 4: Handle player input. In every game iteration, a player must input his move. Which box? : ", end="") move = int...